Remove keyType assert in getKeyType()
Summary: We've talked about removing this assert in the past and I think it's
finally time for it to go. It's caught some real bugs but we're in a much
better place wrt weird types now. In the case of the new test I added, keyType
was {Int|Dbl} on ARM, since the AddO was getting interpOned.

Reviewed By: @edwinsmith

Differential Revision: D1570728