Handle multiple pass-through instructions when merging local values
Summary: FrameState::merge had logic to keep the local value if one of the
branches had a value that was the dest of a pass-through instruction
whose source came from another branch.  This diff generalizes that to
handle cases where there are multiple pass-through instructions in one
of the branches.

Reviewed By: @swtaarrs

Differential Revision: D1569596