Close a memory leak in hh_server and hh_ide.js
Summary: When we want to get the errors for something, we use Errors.try or Errors.do. However, not all callers of functions that add errors are guaranteed to be within Errors.try or Errors.do (specifically, dev tools stuff doesn't always do this).

This made a list of errors accumulate within the errors module forever, with no way to get to it.

This fix makes it so that if we're not within errors.try, errors.ignore or errors.do, we crash if we add an error instead of leaking them.

Reviewed By: @jwatzman

Differential Revision: D1571225