Remove the implicit label hack in vasm-xls
Summary: The numbering scheme for positions requires that the first
instruction in a block not have uses, so that we don't end
up with empty lifetime ranges. In XLS, this was acheived by
inserting a DefLabel if the block didn't have one and it
started with an instruction with 1+ uses. In vasm-xls, I
hacked this by leaving an empty position there and
compenstating for it everywhere. But wrongly, because
phidef should be exactly at the block start.

This diff inserts nop instructions similar to the way XLS
originally inserted DefLabel, if a block started with an
instruction with uses. Inserting in vasm requires memcpy,
but blocks tend to be very small on average. I'll measure
JIT speed just in case.

Reviewed By: @swtaarrs

Differential Revision: D1582946