Remove "analyze" target from hphp (aka "Nemo", or "CodeError.js")
Summary: We're going to delete much of the code that the analyze target is
based on, since it's not used for anything else anymore, and we have Hack as a
better alternative for most use cases now.

This is a commit that mostly only removes the build target so it's easy to see
what the last rev that supported Nemo was.

Reviewed By: @elgenie

Differential Revision: D1574985