Add delay timing knobs to throughput performance benchmark driver.
Change the plumbing so that the $options are pushed around more uniformly
into the workers' constructors, and use properties from the $options
directly.

Add various knobs to control delays before/after most sub-process
launches, delays waiting for responses from the server, delays on the
fcgi channel, and the like. These knobs are reflected as command line
arguments to perf.php. The chosen default values seem plausible or
are historical.

These knobs have proven necessary to support slower systems, or a hhvm
UUT that is itself compiled for debugging or coverage, or is profiling
the PHP being executed.

Add knobs to:
  * override the temporary directory name;
  * plumb subprocess output to files, as using pipes leads to deadlock;
  * trace the subprocess invocations.

If these delay knobs are used, then issue a warning message that
the timing results are unsuitable for benchmark comparisons.

Make changes for hack style guide:
  * add , at end of all formal parameter lists, and most actual paramete lists
  * use 80 character lines
  * normalize spacing
  * use 2-space indent for arguments-on-following-lines, rather than cozy style
  * use !== and === preferentially over != and ==

This is a resubmission from scratch, due to a rebase screw up.
This was PR formerly known as:
  https://github.com/facebook/hhvm/pull/3733
  https://reviews.facebook.net/D23265