Skip translation of region blocks that turn out to be unreachable
Summary: Region formers may create regions that contain blocks that the IR can
later prove they won't be reached. When this happens, the region
translator will try to call IRBuilder::startBlock() on a block for
which no FrameState is saved.  This isn't supported: IRBuilder assumes
that, before emitting code to a block, at least one of the
predecessors has been finished.  This diff changes the region
translator to skip blocks that turn out to be unreachable, and it
strengthens the region checker to check both the assumption above and
that a region is in topological order unless loops are enabled.

Reviewed By: @alexmalyshev

Differential Revision: D1578127