Don't crash on uncounted property initializers with AttrDeepInit set
Summary: The property initialization expression in the new test was causing
EmitterVisitor::requiresDeepInit to return true. This is bad for performance
because the value doesn't really need deep initialization, but more
importantly, we had code in the runtime that was assuming AttrDeepInit
properties were always refcounted. The change in emitter.cpp fixes the first
problem, which might help perf a little, and the change in object-data.cpp
protects us against any other expressions that could produce an uncounted
value.

Reviewed By: @markw65

Differential Revision: D1587716