Add build-root option to hhvm_wrapper
Summary: Being able to tell hhvm_wrapper which build root is handywhen you
have multiple builds in separate directories. Currently hhvm_wrapper
assumes a fixed location for its executables it depends on (hphp and
hhvm). This patch adds an optional command line option to specify the
build root and threads it through to the various places that need it.

Reviewed By: @jdelong

Differential Revision: D1586515