Rename ir.h to ir-opcode.h
Summary: ir.h was starting to turn into a central, monolithic header for the
IR. It used to hold structs for exceptions and guard relaxation.  Other
files were using it to indirectly include (and use) translator-runtime,
physreg, and the x64 ABI.

This renames it to a more appropriate name for what it does, it's just meant
to be a header for the IR Opcode type and related macros.

Reviewed By: @edwinsmith

Differential Revision: D1585361