Fix hphpd crash
Summary: This is a better version of D1591886, which was
release only. It moves the breakpoint filters into
the requestInjectionData, because we have maps
from various things to ThreadInfo, but no maps to
the ExecutionContext.

It makes the DebugHookHandlers singletons, so that
we don't have to delete them racily.

It also fixes the debugger's EventHooks, so that we set/clear
them on the correct threads.

Reviewed By: @jdelong

Differential Revision: D1593790