Apply negation to cgCoerceType's use of emitTypeTest()
Summary: emitTypeTest() expects the caller to negate its
test results when jumping.  cgCoerceStk currently violates
this on the promise that it doesn't use specialized types.

This diff teaches cgCoerceStk to behave nicely by negating the
condition and jumping around multiple blocks.

Vasm then cleans up this dance and we wind up with the same x86
as we're currently generating.

Reviewed By: @swtaarrs

Differential Revision: D1588928