Restore Instr_TC stat counter
Summary: I'm a little concerned that this is going to result in a bunch of extra
work for the sake of accuracy. The old counter (plus an alternate
implementation that I tried) would overcount in the presence of side
exits, because it incremented the counter in bulk at the beginning of
each block. I guess this is OK though since it's only going to result in
extra code if stats are enabled.

Reviewed By: @swtaarrs

Differential Revision: D1594333