Don't propagate states past noreturn functions or Awaits that throw
Summary: Outside of FPI regions this is pretty easy to do.  Inside of them, the
restrictions on bytecode for FPI regions make this easiest to do by not
eliminating the unreachable code---this means running the abstract interpreter
on it, but we don't want any types it infers to leak into successor blocks.
Handle this by adding a new 'unreachable' flag for states and setting up the
state merge rules to handle it appropriately.  When the optimizer sees
unreachable code, it tries to remove it only if it isn't in an FPI region.

Also, change the cases where this happens to push TBottom, and fix a few bugs
where subtypeOf was being used when we really meant something else.
(TBottom.strictSubtypeOf(TObj) is true, but that doesn't mean it has specific
information about an object.)  There are other cases where we'll do things like
push TTrue if a TBottom is being tested for is_int---these are all harmless
because the code is unreachable if you have a TBottom involved.

Reviewed By: @jano

Differential Revision: D1596487