Fixes edge case involving bool args
Summary: If it's possible that the result of a memoize call can be null, we emit an array_key_exists check to see if the result is already cached. For boolean arguments, it would warn and always return false. So this casts the boolean arg to make the scenario work properly.

In order to make the code a little easier, this diff takes ?bool out of the fast path and makes it use the serialiation function. Since I can't actually find a use of ?bool as a typehint in tbgs, I think it's a reasonable choice.

Reviewed By: @elgenie, @jano

Differential Revision: D1598744