Fix symbolizing of OOM stack when text is mapped huge
Summary: When we map out part of .text and re-map it with huge pages, we lose
the address range -> filename mapping in /proc/self/maps, which
folly::symbolizer uses to figure out which file to search for symbols.  This
adds a pass to symbolize using /proc/self/exe, assuming no relocation.

Reviewed By: @markw65

Differential Revision: D1580689