Add an ini_get_all test which gets everything
Summary: In order to ensure we are not landing breaking changes to config options, particularly
to names of those options, let's have a test that simply outputs all of the options
and their values when the test runner runs.

This test will only work with a vanilla, non-repo mode test run, so we have a `.norepo` file for this test.

Broke this up into three test parts to avoid pcre limitations when regex-ing over one big file.

Reviewed By: @ptarjan

Differential Revision: D1591831