Implement jumps & more in llvm backend
Summary: This diff implements enough of the llvm backend to emit the basic
shape of an entire vasm unit (but not all the contents) in my tiny test
program. All the jumps are there, along with enough other stubbed out
operations to pass the llvm verifier and make it through codegen. Since we
don't yet plumb through vmfp and vmsp, most loads are nonsense. The resulting
code from llvm is garbage, of course, so we still punt back to the x64 backend
after the end. Some of the higher-level changes not strictly related to jumps
are:

- When we see a non-virtual Vreg, we ignore it or substitute in some bogus
  value. We'll eventually need to get to a place where there are no physical
  Vregs in the vasm unit, then this code can go away.
- Install an llvm error handler so fatal errors go through our normal
  stacktrace pipeline.
- Run the llvm module verifier after emitting code for the unit.
- Add a landingpad vasm instruction, corresponding to the llvm landingpad
  instruction. It marks the beginning of a catch region and is somewhat
  analogous to the hhir BeginCatch instruction.

Reviewed By: @maksfb

Differential Revision: D1608812