Stop punting on '' === $s
Summary: Codegen was expecting simplifier to have canonicalized constants to the
right, but simplifier didn't do so in the specific case of Same on
strings. Thanks to vasm, it makes no difference to codegen whether
constants are canonicalized right, so just remove the check from
codegen.

Found this by looking for InterpOne while running Wordpress. This takes
it from 0.3-0.5% of bytecodes interp-oned on most endpoints down to a
consistent 0.1%.

Reviewed By: @ottoni

Differential Revision: D1603546