Fix test/slow/ext_gd/imageweb.php
Summary: @​rrh made me aware that the test was not passing any more with a more recent version of libvpx.
This commit corrects the test and should be robust enough for any future update.

Instead of just printing the header and the file size, it now checks if the header is valid and the specified length matches the real length of the file.

Fixes #3927
Closes https://github.com/facebook/hhvm/pull/3929

Reviewed By: @JoelMarcey

Differential Revision: D1603569