Fix annoying typechecker tests that have to be updated with every single change
Summary: I didn't know that hh_show works that way when I wrote them. I switch them to instead pass the value to functions and use the lack of error message as a verifier. It's not checking it as strictly as comparing hh_show, but the current approach is not acceptable.

Reviewed By: @jwatzman

Differential Revision: D1606478