Initial support for handling builtins more intelligently in hhbbc
Summary: This supports constant folding them from a whitelist, and adds a place
to write specific handling for builtins at the type level.  (Only one of these
is implemented for now, get_class, because it looked pretty trivial for a test
case (this diff cuts about 20% of the calls to it out).)

For constant folding, this requires the ability to change a C-flavor'd stack
element into an R, which is weird, but this adds it.  Followup diff will
peephole those away.

Reviewed By: @bertmaher

Differential Revision: D1612167