Unroll AKExists into a series of comparisons on small static arrays
Summary: This is meant to go hand-in-hand with my other change to transform calls
to `in_array` into AKExists, though of course this will benefit organic
uses of AKExists too.

The CFG this generates is suboptimal. It generates a series of diamonds
(one per element of the haystack array) whereas it would ideally result
in something like this:

```
B1 --F--> B2 --F--> B3 --F--+
|         |         |       |
T         T         T       |
|         |         |       |
v         v         v       v
             Bdone
```

I'd have to expose additional API in IRBuilder to do this (since the
HhbcTranslator code would have to create and append its own Blocks) and
I wasn't sure it would be worthwhile. The worst that will happen is a
few redundant unconditional jumps, which we should actually be able to
optimize away downstream.

Reviewed By: @edwinsmith

Differential Revision: D1599893