Clean up MainInit.go.
Summary: It seems like there's no need to run ServerPeriodical.init when
we are doing the one-shot hh_server --check. If so, we can easily factor
the four statements into one if-then block.

I will be refactoring the init process in subsequent commits in preparing for
the save / restore state feature, and this change helps that refactor.

Reviewed By: @jwatzman

Differential Revision: D1610704