FIXME applies to the line of the next lexer token
Summary: This means you have multiple FIXMEs on their own line, and they
will apply to the first non-comment thing after them. Requires a bit of
ugly in the lexer -- it's unclear to me that at this point we shouldn't
just move this over to the parser, getting scary close to actually
parsing, but this is easiest for now.

Reviewed By: @elgenie

Differential Revision: D1614498