Deal with double nulls in an even smarter way
Summary: `Typing.non_null` attempts to deal with double nulls, but
doesn't quite do the entire job, as comes up a lot with `idx2`, see new
test case. This threads information thru `unify` and `sub_type` to deal
with that, along with starting to set up a type `apply` function which
we may eventually want to write. (I originally started trying to write
such a function, which is why this took me so long to crank out, but
ended up just doing this simpler approach.)

Reviewed By: @int3, @elgenie

Differential Revision: D1603190