De-hhvmify hhmeasure
Summary: It'd be nice to be able to use hhmeasure as a tool to comparatively
microbenchmark hhvm and other php implementations, so let's get rid of
the hhvm-specific language extensions.

I considered inlining the implementation of idx() at the callsites,
instead of defining my own. I went with this because I have a hunch that
a lot of real codebases probably define their own idx-like thing, so
this is somewhat realistic. Any thoughts?

Reviewed By: @edwinsmith

Differential Revision: D1614251