Build against bundled sqlite is 3.8
Running hhvm against sqlite 3.8 fails with:
'COMMIT;' --> (5) cannot commit transaction - SQL statements in progress

Without this patch the cmake scripts will use the bundled copy of sqlite
if 3.8.4 <= version < 3.8.6.  With this patch the range is now
3.8.0 <= version.

Cleans up the output CMake spits out when it 3.8 to make it more clear.

Closes #3994