Better handling of RequestEventHandlers
Summary: Its possible for a RequestEventHandler to get re-initialized
after its requestShutdown method has been called.

At the end of the request, we destroy the ExecutionContext, and so we
are left with a RequestEventHandler whose inited flag is set, but which
is not in the ExecutionContext. As a result, we never call requestInit
or requestShutdown for that handler ever again.

This tries to prevent the obvious ways that this can happen, and checks
for it when destroying the ExecutionContext.

Reviewed By: @ptarjan

Differential Revision: D1623445