Convert resume and retransopt to vasm
Summary: resume and retransopt were simply macro operations that do a
service request by preparing a number of argumets, saving
vmsp and vmfp, then returning to enterTCHelper with a service
request opcode.

Add a RegMask argument to the ret{} instruction so we can exactly
model the lifetime of the return-value registers and service-request
argument registers.

This only ports enough of the old emitServiceReqWork functions to
support what was needed for resume and retransopt. Adding support
for ephemeral service requests will take a bit more fleshing out.

Reviewed By: @swtaarrs

Differential Revision: D1622304