Abstract out platform dependent build steps
Summary: Previously the platform-specific steps were inline in the
benchy script. This patch abstracts that part out to a Platform object
for which any platform can provide an implementation. benchy then
decides which Platform object to instantiate based on the contents of
the .benchy file.

This patch also adds shebangs to the various scripts so that they can be
run without invoking python directly. This also enables the shell tab
completion to work correctly for branches.

Reviewed By: @jdelong

Differential Revision: D1620328