Remove some uses of KindOfInvalid
Summary: This takes care of three easily-removed sorts of uses of KindOfInvalid.

The first is dead code; this includes:
  - Type::getDataType() and Type::getHhvmDataType() in the compiler
  - getArgType() in ext/extension.h

The second is easy, file-local conversions to MaybeDataType; these
include:
  - EmitterVisitor::analyzeSwitch()
  - get_option_value_type() in ext_mysqli
  - convert_for_pow() in ext_math

The last is cases where we can't convert to MaybeDataType because the
DataType is part of a TypedValue.  For these, we introduce
kInvalidDataType, which is identical to KindOfInvalid but is not part of
the DataType enum, so that we can expunge KindOfInvalid from case
statements when the remaining uses are deleted.  These include:
  - logic around `mainReturn' in the compiler
  - MixedArray
  - HashCollection
  - NameValueTable

Finally, one miscellaneous s/KindOfInvalid/kInvalidDataType was made:
  - The DataType serialized in `constant->valueText' in
    ClassInfoUnique(); not sure where this comes from.

Reviewed By: @jdelong

Differential Revision: D1624839