Convert most fssa() instructions to make call-target decisions at codegen time
Summary: We've had a long-standing issue since we do the indexing in the
optables at translation time, which means that changes (refinments) in the
instruction arg types in later optimization passes can break things.  This came
up again recently (and hopefully will come up more in the future), so let's fix
it.  For most cases this just moves the optable indexing to codegen time---but
there's were also a decent number of opcodes that used fssa() for no apparent
reason (or in some cases used optabs just to handle a bool, which seems a bit
excessive), so those were just converted into if statements or normal
native-calls on case-by-case basis.

Reviewed By: @swtaarrs

Differential Revision: D1625316