Resolve issue #3812
The 'defaultText' field inside the 'info' array of a ReflectionParameter
instance is matched against a regex that checks if it complies with the
accepted format for labels and constants names.

Then it checks whether or not it is one of the following reserved words:
'NULL', 'false' and 'true'. If it is, then the function returns FALSE.
Otherwise, TRUE.