Fix some emitPropSpecialized code relating to intermediate property accesses
Summary: It turns out that this code doesn't seem to have been running anymore,
since we weren't getting any appropriate knownCls values through to
getPropertyOffset for intermediate prop operations on nullable object types.
Also, this tweaks needFirstRatchet to avoid some unnecessary control flow on
the first prop dim even if we don't know the type of the property (but know the
type of the base).

Reviewed By: @alexmalyshev

Differential Revision: D1625428