Have FrameState track in and out states of blocks.
Summary: Currently FrameState uses its map of snapshots to map blocks to their
in-states, except for when we're building, then we map blocks that have
already been processed to their out-state for
IRBuilder::insertLocalPhis.

To clean things up and reduce confusion, lets just store both the in and
out state for each block.

Reviewed By: @swtaarrs

Differential Revision: D1619253