LZ4_uncompress is obsolete, breaks build with out-of-tree LZ4.
Summary: LZ4_uncompress() has been replaced 1:1 with LZ4_decompress_fast().  When
you use an out-of-tree liblz4 r122 or beyond, LZ4_uncompress doesn't exist
anymore.  From bottom of lz4.h:

 [...]
 /*
 Obsolete decompression functions
 These function names are deprecated and should no longer be used.
 They are only provided here for compatibility with older user programs.
 - LZ4_uncompress is the same as LZ4_decompress_fast
 - LZ4_uncompress_unknownOutputSize is the same as LZ4_decompress_safe
 */
 int LZ4_uncompress (const char* source, char* dest, int outputSize);
 [...]

LZ4_decompress_fast() is already present in third-party's copy, so even if this
is not breaking a build with an internal lz4, it is for Debian and any other that
may have the library outside of the HHVM tree detected by CMake.

Closes: https://github.com/facebook/hhvm/issues/3610

Reviewed By: @ptarjan

Differential Revision: D1628566