Some cleanup during simplifier audit; fix one bug in xorTrueImpl
Summary: Some clean up while auditing simplifier instructions for #5394328.  It
looks like the only incorrect cases are the ones I've added TODO's to (all
comparison op related).  Conditional jump folding already explicitly handled
reference counted values (with a comment explaining this exact case---blames
back to the original code).  The only functional change in this diff is to fix
a bug in xorTrueImpl for ColIsEmpty and ColIsNEmpty.

Reviewed By: @alexmalyshev

Differential Revision: D1630229