Fix remaining simplifier bug; try to add some validation assertions
Summary: This turned out to be way less of a general problem than we were
worried.  Almost all the simplifier rules just do constant folding or try to
eliminate no-op opcodes.  Only two were doing this wrong, both in the
not-folding stuff (other one is fixed in previous diff).  I audited all the
cmpImpl code and turns out to be fine.

This diff tries to add an assertion pass (which at least should document the
rules), but it's a little hokey for our stack opcodes: they need to be
special-cased because the simplifier can't easily double-check that the values
are known to be available due to being on the eval stack, but we know this is
true (in part because of bytecode invariants).  The requirements for those guys
to be correct is a little indirect (partially follows from our rules about
StkPtr use)---maybe I should add some documentation about it?

Reviewed By: @alexmalyshev

Differential Revision: D1630802