Split hhbc-translator.cpp up into several files
Summary: Move several large chunks of it into different files.  The main chunk
is still a mess (and the largest one still by a factor of 3), but this diff
gets the same build time win from D1625575 without being so hacky about it.
Most of what's left is either small bytecodes (String, Int, AddNewElem etc) or
"scaffolding" used by the whole thing, which is probably another logical split
(I don't know what to call the scaffolding though).  However, I'd like to take
this as is so far---making it more split up than this is worth doing, but it
won't get any immediate build time improvements because code-gen-x64 is the
long tail now (by 3 seconds).

Reviewed By: @alexmalyshev, @swtaarrs

Differential Revision: D1631192